+++ poem/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# Unreleased
+# [1.0.7] 2010-10-21
+
+- Update some docs.
+
+# [1.0.6] 2010-10-20
 
 - `Cors` middleware allows all HTTP methods and headers by default.
 - Add `Cors::allow_origins_fn` method.